+++ elastic-job-cloud/elastic-job-cloud-scheduler/src/main/java/com/dangdang/ddframe/job/cloud/scheduler/env/BootstrapEnvironment.java
@@ -58,7 +58,19 @@ public final class BootstrapEnvironment {
         } catch (final IOException ex) {
             log.warn("Cannot found conf/elastic-job-cloud-scheduler.properties, use default value now.");
         }
+        setPropertiesByEnv(result);
         return result;
+    }
+    
+    private void setPropertiesByEnv(final Properties prop) {
+        for (EnvironmentArgument each : EnvironmentArgument.values()) {
+            String key = each.getKey();
+            String value = System.getenv(key);
+            if (!Strings.isNullOrEmpty(value)) {
+                log.info("Load property %s with value %s from ENV.", key, value);
+                prop.setProperty(each.getKey(), value);
+            }
+        }
     }
     
     /**